client/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
isodate==0.6.1
lxml==5.2.0
mysqlclient==2.2.4
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
wheel==0.41.2
zeep==4.2.1
